Wayを別クラス
Wayを別クラス
