+++ Assets/MainContent.cs
@@ -1,52 +1,24 @@
 ﻿using UnityEngine;
 using System.Collections;
 using System.Collections.Generic;
-//みち
-internal class WayPoint{
-	internal Vector3 pos;//
-	
-	internal Room Room;
-	internal List<WayPoint> link = new List<WayPoint>();
-	internal List<GameObject> model = new List<GameObject>();//自身から次への道
 
-	internal void AddWay(WayPoint point){
-		//中継必要？
-		if (pos.x != point.pos.x && pos.z != point.pos.z) {
-			var midPoint = new WayPoint ();
-			if(Random.Range(0,2) == 0){
-				midPoint.pos.x = pos.x;
-				midPoint.pos.z = point.pos.z;
-			}else{
-				midPoint.pos.x = point.pos.x;
-				midPoint.pos.z = pos.z;
-			}
-			AddWay (midPoint);
-			midPoint.AddWay (point);
-			return;
-		} else {
-			link.Add (point);
-			point.link.Add (this);
-		}
-	}
-	internal void createModel(MainContent parent){
-		var wayModel = (GameObject)GameObject.Instantiate(parent.WayPrefab, pos, Quaternion.identity);
-		wayModel.name = "WayPoint";
-		model.Add(wayModel);
-		foreach(var way in link){
-			if(way.model.Count > 0){
-				continue;//作成済み
-			}
-			//みち
-			wayModel = (GameObject)GameObject.Instantiate(parent.WayPrefab, (way.pos + pos)/2, Quaternion.LookRotation(way.pos - pos));
-			var modPos = way.pos - pos;
-			wayModel.transform.localScale = new Vector3(1,1, modPos.magnitude-1);
-			wayModel.name = "Way";
-			model.Add(wayModel);
+/*
+O:Room
+P:WayPoint
+W:Way
 
-			way.createModel(parent);
-		}
-	}
-}
+
+OOO     OOO
+OPO     OOP
+OOO     OOO
+ W        W
+ W        W
+ W        W
+OOO       W
+OPOWWWWWWWP
+OOO
+ */
+
 public class MainContent : MonoBehaviour {
 
 	//定義===================================
@@ -158,4 +130,13 @@ public class MainContent : MonoBehaviour {
 		PlayerCharacter.transform.position = Rooms [0].way.pos;
 		
 	}
+
+
+	//マップ移動を検証
+	internal void CheckMapCollision (Character target) {
+	}
+	//マップ移動を検証
+	internal void CheckObjCollision (Character target) {
+	}
+
 }
