+++ Assets/WayPoint.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 5aa235e1633ffce48acde863053e6620
+timeCreated: 1436164248
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
